% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chamber_arduino.R
\name{chamber_arduino}
\alias{chamber_arduino}
\title{FUnction to calculate CO2 flux from arduino measurements}
\usage{
chamber_arduino(datelim, data = NULL, return_ls = F)
}
\arguments{
\item{datelim}{ymd_hm}

\item{data}{default = NULL data is loaded from inj_arduino_pfad}

\item{return_ls}{F = return flux data.frame if T = returns list with [[1]] = flux and [[2]] data}
}
\value{

}
\description{
FUnction to calculate CO2 flux from arduino measurements
}
