% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_unit.R
\name{change_unit}
\alias{change_unit}
\title{Function to change units of a vector}
\usage{
change_unit(x, unit_in = NULL, unit_out, class_out = "numeric")
}
\arguments{
\item{x}{vector}

\item{unit_in}{unit of input vector as character}

\item{unit_out}{output unit as character}

\item{class_out}{class of output either \code{"numeric"} or \code{"units"}}
}
\value{

}
\description{
Function to change units of a vector
}
