% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check.packages}
\alias{check.packages}
\title{load packages and install them if they are not already installed}
\usage{
check.packages(pkgs)
}
\arguments{
\item{pkgs}{character string with package names}
}
\value{
nothing returned
}
\description{
load packages and install them if they are not already installed
}
\examples{
packages<-c("stringr","lubridate","RSQLite","odbc","reshape2","zoo","ggplot2","ggnewscale","xlsx","readxl")

check.packages(packages)
}
