% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comsol_exe.R
\name{comsol_exe}
\alias{comsol_exe}
\title{Function to execute Comsol Modell with defined parameters}
\usage{
comsol_exe(
  modelname,
  input_pars = NULL,
  outfile_new = NULL,
  outfile_raw = "CO2_flux_prod.txt",
  job = "b1",
  overwrite = F
)
}
\arguments{
\item{modelname}{name of the Comsol Modell without .mph ending}

\item{input_pars}{named vector or dataframe where names represent Parameter names as defined in Comsol}

\item{outfile_name}{name that the Comsol Export file should get}
}
\value{

}
\description{
Function to execute Comsol Modell with defined parameters
}
\examples{
comsol_exe(model="Produktionseimer",input_pars=input_pars_9,outfile="CO2_flux_prod_9.txt")
}
