% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comsol_exe_nruns.R
\name{comsol_exe_nruns}
\alias{comsol_exe_nruns}
\title{Function to execute Comsol Modell with defined parameters}
\usage{
comsol_exe_nruns(
  modelname,
  data_list,
  outfile_names,
  outfile_raw = "CO2_optim.txt",
  COMSOL_progammpath = COMSOL_progammpfad,
  COMSOL_exepath = COMSOL_exepfad,
  job = "b1",
  overwrite = F,
  nruns = 10
)
}
\arguments{
\item{modelname}{name of the Comsol Modell without .mph ending}

\item{outfile_raw}{name that COMSOL gives the outputfile}

\item{COMSOL_progammpath}{path to COMSOL project file}

\item{COMSOL_exepath}{path to comsolbatch.exe}

\item{job}{name of the job that is defined in the COMSOL project file}

\item{overwrite}{should existing modellresults be overwritten or not}
}
\value{

}
\description{
Function to execute Comsol Modell with defined parameters
}
