% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{D0_T_p}
\alias{D0_T_p}
\title{Function to calculate temperature dependent D0}
\usage{
D0_T_p(T_C = 20, gas = "CO2", p_kPa = 101.325)
}
\arguments{
\item{T_C}{Temperature in Celsius}

\item{gas}{CO2 or CH4}

\item{p_kPa}{pressure in kPa}
}
\value{

}
\description{
Function to calculate temperature dependent D0
}
\examples{
D0_temp(15)
}
