% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ms_as_int.R
\name{date_ms_as_int}
\alias{date_ms_as_int}
\title{Function to convert a date with 10Hz to Integer}
\usage{
date_ms_as_int(date)
}
\arguments{
\item{date}{input date with 10 Hz}
}
\value{
date_int as Interger in 100ms since 2022-01-27 00:00:00.0
}
\description{
Function to convert a date with 10Hz to Integer
}
