% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_time.R
\name{diff_time}
\alias{diff_time}
\title{function to make difftime easier}
\usage{
diff_time(time, unit = "secs", out_class = "numeric", ...)
}
\arguments{
\item{time}{datetime vector}

\item{unit}{character}

\item{out_class}{class of output}

\item{...}{other params passed to difftime}
}
\value{

}
\description{
function to make difftime easier
}
