% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectionrate.R
\name{injectionrate}
\alias{injectionrate}
\title{Function to calculate Injectionrate for specific Diffusionchamber measurement}
\usage{
injectionrate(
  datelim,
  Pumpstufen,
  group = "Pumpstufe",
  closing_before = 20,
  closing_after = 20,
  opening_before = 0,
  opening_after = 10,
  t_max = 6,
  t_init = 2,
  t_min = 3,
  ...
)
}
\arguments{
\item{datelim}{time intervall}

\item{Pumpstufen}{verwendete Pumpstufen bei den unterschiedlichen Versuchen}

\item{group}{name of the column that should be used for grouping results}

\item{...}{#other parameters parsed to calc_flux}
}
\value{

}
\description{
Function to calculate Injectionrate for specific Diffusionchamber measurement
}
\examples{
split <- injectionrate(datelim = datelim,Pumpstufen = c(1:5,5,rep(NA,4),1:4,1:5),group="Pumpstufe",
closing_before = 20,
closing_after = 20,
opening_before = 0,
opening_after = 10,
t_max=6,
t_init = 1,
t_min=3)
}
