% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_leave_NA_time.R
\name{leave_NAtime_plot}
\alias{leave_NAtime_plot}
\title{Title}
\usage{
leave_NAtime_plot(
  y = "CO2",
  col = "tiefe",
  group = "Pumpstufe",
  data,
  timestep = 10,
  plot = T,
  ylab = "CO2 [ppm]"
)
}
\arguments{
\item{ylab}{}
}
\value{

}
\description{
Title
}
