% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{ppm_to_mol}
\alias{ppm_to_mol}
\title{Function to convert units from ppm to mol / m^3 or reverse}
\usage{
ppm_to_mol(
  data,
  unit_in = "ppm",
  out_class = "numeric",
  p_kPa = 101.3,
  T_C = 20
)
}
\arguments{
\item{data}{vector with values that should be converted from ppm to mol or reverse}

\item{unit_in}{unit of input data if this is ppm output will be in mol / m^3 and vice versa}

\item{out_class}{class of the output as character eg. \code{"numeric"} (default) or \code{"units"}}

\item{p_kPa}{pressure in kPa}

\item{T_C}{Temperature in °C}
}
\value{
vector with converted units
}
\description{
Function to convert units from ppm to mol / m^3 or reverse
}
\examples{
ppm_to_mol(CO2_ppm,"ppm")
ppm_to_mol(0.0168,"mol/m^3")
}
