% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GGA.R
\name{read_GGA}
\alias{read_GGA}
\title{Function to get data from a sql database}
\usage{
read_GGA(
  db.name = "GGA.db",
  table.name = "micro",
  datelim = NULL,
  cols = "*",
  sqlpath = sqlpfad,
  ggapath = ggapfad
)
}
\arguments{
\item{db.name}{name of the database
for "dynament.db" either \code{"dynament_test"} or \code{"samplerx"}
for "GGA.db" either \code{"gga"} or \code{"micro"}}

\item{table.name}{name of the table in the database}

\item{datelim}{limits of the time period that is loaded as character or POSIXct if all Data should be loaded \code{datelim = NULL}}

\item{cols}{names of the colums that will be loaded. default is all colums: \code{cols = "*"}}

\item{sqlpath}{path to the database}
}
\value{
data.frame
}
\description{
get data from a sql database. A specific time period can be selected
}
\examples{
#Zeitrahmen festlegen
datelim<-c("2020-01-15 10:00:00","2020-01-21 10:00:00")

#read_db anwenden
dynament_raw<-read_db("dynament.db","dynament_test",datelim,korrektur_dyn=F)
GGA<-read_db("GGA.db","micro",datelim,"CO2,CO2dry")
}
