% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sampler.R
\name{read_sampler}
\alias{read_sampler}
\title{Read sampler data from database}
\usage{
read_sampler(table.name = "sampler1u2", format = "long", ...)
}
\arguments{
\item{table.name}{name of the table in the database}

\item{format}{long or wide}

\item{...}{other parameters parsed to read_db such as datelim or cols
(see help of read_db)}
}
\value{

}
\description{
Read sampler data from database
}
\examples{
datelim<-c("2020.02.19 12:05:00","2020.02.20 09:35:00")
data <- read_sampler("sampler1",datelim = datelim)
}
