% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_grid.R
\name{rel_grid}
\alias{rel_grid}
\title{Title}
\usage{
rel_grid(n = 0.1, ...)
}
\arguments{
\item{n}{distance of grid lines number between 0 and 1}

\item{...}{geom_path parameters such as colour, alpha, size, etc.}
}
\value{

}
\description{
Title
}
\examples{
cowplot::ggdraw()+
cowplot::draw_plot(plot)+
rel_grid(0.01,color="grey")+
rel_grid(0.1)
}
