% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_file_from_db.R
\name{rm_file_from_db}
\alias{rm_file_from_db}
\title{Function to remove a specific file from the Database}
\usage{
rm_file_from_db(
  filename,
  table_name = "sampler1",
  db_name = "dynament.db",
  path = dynpfad
)
}
\arguments{
\item{filename}{name of file}

\item{table_name}{name of table in db}

\item{db_name}{name of database}

\item{path}{path to file}
}
\value{

}
\description{
Function to remove a specific file from the Database
}
\examples{
rm_file_from_db("^b0_work_20200526_102154_smp2.csv")
}
