% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Compute RootMeanSquareError}
\usage{
RMSE(mod, obs, normalize = F)
}
\arguments{
\item{mod}{modelled data}

\item{obs}{observed data}
}
\value{
RMSE
}
\description{
Compute RootMeanSquareError
}
\examples{
RMSE(CO2_mod,CO2_obs)
}
