% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comsol_nruns.R
\name{run_comsol_nruns}
\alias{run_comsol_nruns}
\title{Function to}
\usage{
run_comsol_nruns(
  data = data,
  mod_dates,
  overwrite = F,
  read_all = T,
  n_DS = 3,
  plot = F,
  offset_method = "drift",
  z_soil_cm = 150,
  modelname = "Diffusion_freeSoil_anisotropy_optim_3DS_50runs",
  file_suffix = NULL,
  comsolpath = comsolpfad,
  long = T,
  ...
)
}
\arguments{
\item{data}{dataframe mit Messwerten}

\item{mod_dates}{Zeitpunkte zu denen modellierung statt finden soll}

\item{overwrite}{sollen existierende Modelloutputs überschrieben werden}

\item{read_all}{sollen alle Modellergebnisse der gewählten methode geladen werden}

\item{n_DS}{Anzahl an DS Schichten}

\item{plot}{sollen die Ergebnisse geplotten werden}

\item{offset_method}{name der methode zur offset korrektur \code{"gam"} oder \code{"drift" ...}}

\item{z_soil_cm}{Tiefe des Modells in cm}

\item{modelname}{name des COMSOL Projekts}

\item{file_suffix}{Dateiendung zB. um alternative Modellläufe zu kennzeichnen}

\item{comsolpath}{}
}
\value{

}
\description{
Function to
}
