% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comsol_old.R
\name{run_comsol_old}
\alias{run_comsol_old}
\title{Title}
\usage{
run_comsol_old(
  data = data,
  mod_dates,
  overwrite = F,
  read_all = T,
  n_DS = 3,
  plot = F,
  offset_method = "gam",
  optim_method = "snopt",
  z_soil_cm = 150,
  modelname = "Diffusion_freeSoil_optim_3DS"
)
}
\arguments{
\item{modelname}{}
}
\value{

}
\description{
Title
}
