% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chamber.R
\name{split_chamber}
\alias{split_chamber}
\title{Function to automatically identify closing and opening times of chamber measurents}
\usage{
split_chamber(
  data,
  closing_lim = 30,
  opening_lim = -10,
  t_max = 3,
  t_init = 0.1,
  t_min = 2,
  gas = "CO2",
  round_intervall = "1 min",
  adj_openings = T
)
}
\arguments{
\item{data}{data.frame with gas concentrations of several chamber measurements}

\item{closing_lim}{maximal concentration-gradient before closing the chamber}

\item{opening_lim}{minimal concentration-gradient before opening the chamber}

\item{t_max}{maximal measurement time (minutes), if time between closing and opening
exceeds this time the rest of the measurement will not be included for the calculation
of the flux}

\item{t_init}{initial time (minutes) after closing of the chamber that will be excluded from the calculations}

\item{t_min}{minimum timespan for each chamber measurement in minutes}

\item{gas}{name of the gas as character}

\item{round_intervall}{}

\item{adj_openings}{logical; if \code{TRUE} opening times will be adjusted
to never be smaller than closing times}
}
\value{
input data.frame with colums messid (number of the measurement) and zeit (time in minutes after closing) added
}
\description{
Function to automatically identify closing and opening times of chamber measurents
}
\examples{
test <- split.chamber(data,
closing_th = 40,
opening_th = -40,
t_max=Inf,
t_init=0,
t_min=5)
}
