% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_daterange.R
\name{sub_daterange}
\alias{sub_daterange}
\title{Function that returns a subset of data within the given daterange}
\usage{
sub_daterange(data, daterange, datecol = "date")
}
\arguments{
\item{data}{}

\item{daterange}{}

\item{datecol}{}
}
\value{

}
\description{
Function that returns a subset of data within the given daterange
}
