% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GGA.db.R
\name{unzip.files}
\alias{unzip.files}
\title{Funktion um alle .zip dateien in einem Ordner incl. Unterordnern zu entpacken}
\usage{
unzip.files(path)
}
\arguments{
\item{path}{the path that the zip files are in}
}
\value{
nothing returned
}
\description{
Alle .zip dateien in einem Ordner incl. Unterordnern werden entpackt und im Unterordner unzipped gespeichert
}
\examples{
unzip.files("H:/FVA-Projekte/P01677_WindWaldMethan/Daten/Urdaten/GGA/")
}
