% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{update_dynament.db}
\alias{update_dynament.db}
\title{update the dynament database}
\usage{
update_dynament.db(
  table.name = "dynament_test",
  path = dynpfad,
  sqlpath = sqlpfad,
  metapath = metapfad_dyn
)
}
\arguments{
\item{table.name}{name of the table in the databese that shall be updated}

\item{path}{the path that the db_log file and the data are in}

\item{sqlpath}{the path to the database}

\item{metapath}{the path to the metadata for correcting the sensor output of a specific sensor id}
}
\value{
nothing returned
}
\description{
all files that are not listed in the db_log file are formated and imported into the dynament database
}
\examples{
update_dynament.db("dynament_test")
}
