% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Em50.R
\name{update_Em50_data.db}
\alias{update_Em50_data.db}
\title{update Em50 database}
\usage{
update_Em50_data.db(table.name = "Em50", path = datapfad_schaui)
}
\value{
nothing returned
}
\description{
adds all the files in the datapfad Em50Data subdirectory to the Em50 database
}
\examples{
Em50.update.db()
}
