% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GGA.db.R
\name{update_GGA.db}
\alias{update_GGA.db}
\title{update the GGA database}
\usage{
update_GGA.db(table.name = c("gga", "micro"), path, sqlpath)
}
\arguments{
\item{path}{the path that the unzipped and the db_log folder are in}

\item{sqlpath}{the path to the database}
}
\value{
nothing returned
}
\description{
all files that are not listed in the db_log folder are formated and imported into the GGA database
}
\examples{
update_GGA.db()
}
