% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_PP.db.R
\name{update_PP.db}
\alias{update_PP.db}
\title{update the PP database}
\usage{
update_PP.db(table.name = "PP_chamber")
}
\arguments{
\item{table.name}{name of the table in the databese that shall be updated}
}
\value{

}
\description{
update the PP database
}
