% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_to_Pa.R
\name{V_to_Pa}
\alias{V_to_Pa}
\title{Function to calculate ressure from the digital output of the Sensors}
\usage{
V_to_Pa(p_adc)
}
\arguments{
\item{p_adc}{}
}
\value{

}
\description{
Function to calculate ressure from the digital output of the Sensors
}
