% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_to_depths.R
\name{variable_to_depths}
\alias{variable_to_depths}
\title{Funktion um messungen einer Variablen im wide Format für andere Messtiefen zu interpolieren}
\usage{
variable_to_depths(variable, colname = NULL, df = data)
}
\arguments{
\item{variable}{name variablen es müssen die spalten im wide format vorliegen mit spaltennamen: "variable_tiefe" z.B. "VWC_-5"}

\item{colname}{name der spalte mit den interpolierten Werten der Variablen}

\item{df}{dataframe}
}
\value{

}
\description{
Funktion um messungen einer Variablen im wide Format für andere Messtiefen zu interpolieren
}
\examples{
data <- variable_to_depths("VWC")
data <- variable_to_depths("T_C","T_soil")
}
