% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechsel_fun.R
\name{wechsel_fun}
\alias{wechsel_fun}
\title{Function to identify periods in sensor calibration}
\usage{
wechsel_fun(
  data,
  col = "CO2",
  ppm_lim = 60,
  min_lim = 10,
  puffer_start = 10,
  puffer_ende = 1
)
}
\arguments{
\item{data}{measured data}

\item{col}{which column should be used}

\item{ppm_lim}{limit for ppm change in same period}

\item{min_lim}{maximal time difference of data points in one period}

\item{puffer_start}{amount of minutes to cut off at the begin}

\item{puffer_ende}{amount of minutes to cut off at the end}
}
\value{

}
\description{
Function to identify periods in sensor calibration
}
\examples{
sampler2_list <- wechsel_fun(data=data_sampler2,col="CO2_Dyn_02")
}
