% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chamber_arduino.R
\name{chamber_arduino}
\alias{chamber_arduino}
\title{FUnction to calculate CO2 flux from arduino measurements}
\usage{
chamber_arduino(
  datelim,
  data = NULL,
  gga_data = T,
  gas = c("CO2", "CO2_GGA", "CH4"),
  t_min = 5,
  t_init = 1,
  t_max = 10,
  t_offset = 0,
  gga = "gga",
  plot = "facets",
  return_ls = T
)
}
\arguments{
\item{datelim}{ymd_hm}

\item{data}{default = NULL data is loaded from inj_arduino_pfad}

\item{return_ls}{F = return flux data.frame if T = returns list with [[1]] = flux and [[2]] data}
}
\value{

}
\description{
FUnction to calculate CO2 flux from arduino measurements
}
