% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{D0_T_p}
\alias{D0_T_p}
\title{Function to calculate temperature dependent D0 in cm2/s}
\usage{
D0_T_p(T_C = 20, p_kPa = 101.325, unit = "cm^2/s", gas = "CO2")
}
\arguments{
\item{T_C}{Temperature in Celsius}

\item{p_kPa}{pressure in kPa}

\item{gas}{CO2 or CH4}
}
\value{

}
\description{
Function to calculate temperature dependent D0 in cm2/s
}
\examples{
D0_temp(15)
}
