% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_daterange.R
\name{daterange_id}
\alias{daterange_id}
\title{Function that returns a vector with the IDs of the rows in data that lie in daterange}
\usage{
daterange_id(data, daterange, datecol = "date")
}
\arguments{
\item{data}{}

\item{daterange}{}

\item{datecol}{}
}
\value{

}
\description{
Function that returns a vector with the IDs of the rows in data that lie in daterange
}
