% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectionrate.R
\name{injectionrate}
\alias{injectionrate}
\title{Function to calculate Injectionrate for specific Diffusionchamber measurement}
\usage{
injectionrate(
  datelim,
  Pumpstufen,
  group = "Pumpstufe",
  closing_lim = 100,
  opening_lim = 0,
  t_max = 6,
  t_init = 1,
  t_min = 2,
  T_C = NA,
  spikes_th = 500,
  difftime_th = 10,
  all_spikes_NA = F,
  data = NULL,
  return_data = F,
  colname = "CO2",
  adj_openings = T,
  ...
)
}
\arguments{
\item{datelim}{time intervall}

\item{Pumpstufen}{verwendete Pumpstufen bei den unterschiedlichen Versuchen}

\item{group}{name of the column that should be used for grouping results}

\item{T_C}{either a numeric value for one Temperature or the name of the Temperture column as character}

\item{return_data}{if T then the function returns a list with the flux as first element and the processed input data as second element}

\item{colname}{name of the CO2 column in the df}

\item{...}{#other parameters parsed to calc_flux}
}
\value{

}
\description{
Function to calculate Injectionrate for specific Diffusionchamber measurement
}
\examples{
split <- injectionrate(datelim = datelim,Pumpstufen = c(1:5,5,rep(NA,4),1:4,1:5),group="Pumpstufe",
closing_before = 20,
closing_after = 20,
opening_before = 0,
opening_after = 10,
t_max=6,
t_init = 1,
t_min=3)
}
