% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_leave_NA_time.R
\name{leave_NAtime_plot}
\alias{leave_NAtime_plot}
\title{Function to plot data with x-axis split into facets between longer periods without data}
\usage{
leave_NAtime_plot(
  y = "CO2",
  col = "tiefe",
  group = "Pumpstufe",
  data,
  timestep = 10,
  ylab = "CO2 [ppm]",
  plot = T,
  adj_grob_size = T,
  ...,
  geom = "line"
)
}
\arguments{
\item{y}{name of the column that should be ploted on y-axis as character}

\item{col}{name of the column that should be used for the colour as character}

\item{group}{name of the column that should be used for finding NA times as character the plot will automatically be split into different facets if the group column is NA for a defined timestep}

\item{data}{data that should be plotted}

\item{timestep}{timestep in minutes that is used to split plot into facets without longer gaps on the time axis}

\item{ylab}{label for y-axis}

\item{plot}{logical indicating whether output is a ggplot object (T) or a data.frame with the extra column \code{"period"} which can be used in facet_wrap to split time-axis}

\item{adj_grob_size}{logical whether grob size should be adjusted}

\item{...}{other parameters that should be parsed to \code{scale_x_datetime(date_breaks = breaks, ...)}}
}
\value{

}
\description{
Function to plot data with x-axis split into facets between longer periods without data
}
\examples{
leave_NAtime_plot(data=data,group="CO2",plot=T)
}
