% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PP.R
\name{read_PP}
\alias{read_PP}
\title{Function to read Pressure data from db}
\usage{
read_PP(datelim = NULL, cols = "*", table.name = "PP_chamber", format = "long")
}
\arguments{
\item{datelim}{}

\item{cols}{}
}
\value{

}
\description{
Function to read Pressure data from db
}
