% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comsol.R
\name{run_comsol}
\alias{run_comsol}
\title{Title}
\usage{
run_comsol(
  data = data,
  mod_dates,
  overwrite = F,
  read_all = T,
  n_DS = 3,
  plot = F,
  offset_method = "gam",
  z_soil_cm = 150,
  modelname = "Diffusion_freeSoil_optim_3DS",
  file_suffix = NULL,
  input_pars = NULL
)
}
\arguments{
\item{modelname}{}
}
\value{

}
\description{
Title
}
