% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dynament.db.R
\name{update_dynament.db}
\alias{update_dynament.db}
\title{update the dynament database}
\usage{
update_dynament.db(table.name = "dynament_test")
}
\arguments{
\item{table.name}{name of the table in the databese that shall be updated}
}
\value{
nothing returned
}
\description{
all files that are not listed in the db_log file are formated and imported into the dynament database
}
\examples{
update_dynament.db("dynament_test")
}
